Changed from dynamic to static linking + ZERO build warnings

+++ Cajo/src/Platform/Windows/WindowsWindow.cpp
@@ -141,7 +141,7 @@ namespace Cajo {
 		{
 			WindowData& data = *(WindowData*)(glfwGetWindowUserPointer(window));
 
-			MouseScrolledEvent event(xOffset, yOffset);
+			MouseScrolledEvent event((float)xOffset, (float)yOffset);
 			data.EventCallback(event);
 		});
